project/src/legal-tales/backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
sentence-transformers==2.2.2
python-dotenv==1.0.0